ui.adapter 패키지 추가
