+++ app/src/main/java/com/example/todaynan/ui/adapter/BoardWriteActivityRVAdpater.kt
@@ -1,14 +1,15 @@
-package com.example.todaynan
+package com.example.todaynan.ui.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import com.example.todaynan.R
 import com.example.todaynan.data.entity.MyLikedPost
 import com.example.todaynan.databinding.MyLikedPostBinding
 
-class BoardReplyActivityRVAdapter(private val items: List<MyLikedPost>) : RecyclerView.Adapter<BoardReplyActivityRVAdapter.ViewHolder>() {
+class BoardWriteActivityRVAdpater(private val items: List<MyLikedPost>) : RecyclerView.Adapter<BoardWriteActivityRVAdpater.ViewHolder>() {
 
     override fun onCreateViewHolder(viewGroup: ViewGroup, viewType: Int): ViewHolder {
         val binding= MyLikedPostBinding.inflate(LayoutInflater.from(viewGroup.context),viewGroup,false)
