+++ app/src/main/java/com/example/todaynan/ui/main/mypage/BoardWriteActivity.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.example.todaynan.BoardLikedRVAdapter
+import com.example.todaynan.ui.adapter.BoardLikedRVAdapter
 import com.example.todaynan.data.entity.MyLikedPost
 import com.example.todaynan.R
 import com.example.todaynan.databinding.ActivityBoardWriteBinding
